+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres _loosely_ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.6] - tbd
+## [0.4.6] - 2026-05-05
 ### Added
 * Added `INode.ReverseInPlace(containment)` and `INode.ReverseInPlace(reference)`.
 ### Fixed